requirements.txt
pymysql
tkcalendar
mysql.connector